+++ Tasks/CodeQuality/PesterIntegration.ps1
@@ -40,7 +40,7 @@ Add-BuildTask -Name PesterIntegration -Jobs SetVariables, {
     $pesterSplat = @{
         Path           = "$BuildRoot\Tests\Integration"
         ModuleManifest = Join-Path $BuildRoot "$script:moduleName.psd1"
-        StreamOutput   = $script:PlumberConfig.StreamPesterOutput
+        StreamOutput   = $script:PlumberConfig.Tasks.PesterIntegration.StreamOutput
     }
     $result = Invoke-PlumberPester @pesterSplat